+++ packages/core/src/Extension.ts
@@ -171,7 +171,7 @@ export class Extension<Options = any, Commands = any> {
     return new Extension<O, C>(config)
   }
 
-  configure(options: Partial<Options>) {
+  configure(options?: Partial<Options>) {
     return Extension
       .create<Options, Commands>(this.config as ExtensionConfig<Options, Commands>)
       .#configure({