+++ packages/core/src/commands/joinForward.ts
@@ -1,0 +1,9 @@
+import { joinForward as originalJoinForward } from 'prosemirror-commands'
+import { Command } from '../types'
+
+/**
+ * Join two nodes forward.
+ */
+export const joinForward = (): Command => ({ state, dispatch }) => {
+  return originalJoinForward(state, dispatch)
+}