+++ packages/core/src/commands/selectNodeBackward.ts
@@ -1,0 +1,9 @@
+import { selectNodeBackward as originalSelectNodeBackward } from 'prosemirror-commands'
+import { Command } from '../types'
+
+/**
+ * Select a node backward.
+ */
+export const selectNodeBackward = (): Command => ({ state, dispatch }) => {
+  return originalSelectNodeBackward(state, dispatch)
+}