+++ packages/html/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/html@2.0.0-alpha.5...@tiptap/html@2.0.0-alpha.6) (2021-01-06)
+
+**Note:** Version bump only for package @tiptap/html
+
+
+
+
+
+# [2.0.0-alpha.5](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/html@2.0.0-alpha.4...@tiptap/html@2.0.0-alpha.5) (2020-12-18)
+
+**Note:** Version bump only for package @tiptap/html
+
+
+
+
+
+# [2.0.0-alpha.4](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/html@2.0.0-alpha.3...@tiptap/html@2.0.0-alpha.4) (2020-12-18)
+
+**Note:** Version bump only for package @tiptap/html
+
+
+
+
+
 # [2.0.0-alpha.3](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/html@2.0.0-alpha.2...@tiptap/html@2.0.0-alpha.3) (2020-12-02)
 
 **Note:** Version bump only for package @tiptap/html