+++ packages/vue-starter-kit/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.9](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/vue-starter-kit@2.0.0-alpha.8...@tiptap/vue-starter-kit@2.0.0-alpha.9) (2021-01-06)
+
+**Note:** Version bump only for package @tiptap/vue-starter-kit
+
+
+
+
+
+# [2.0.0-alpha.8](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/vue-starter-kit@2.0.0-alpha.7...@tiptap/vue-starter-kit@2.0.0-alpha.8) (2020-12-18)
+
+**Note:** Version bump only for package @tiptap/vue-starter-kit
+
+
+
+
+
+# [2.0.0-alpha.7](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/vue-starter-kit@2.0.0-alpha.6...@tiptap/vue-starter-kit@2.0.0-alpha.7) (2020-12-18)
+
+**Note:** Version bump only for package @tiptap/vue-starter-kit
+
+
+
+
+
 # [2.0.0-alpha.6](https://github.com/ueberdosis/tiptap-next/compare/@tiptap/vue-starter-kit@2.0.0-alpha.5...@tiptap/vue-starter-kit@2.0.0-alpha.6) (2020-12-02)
 
 **Note:** Version bump only for package @tiptap/vue-starter-kit